+++ server/db/update_scripts/27_update_v6.12.0_v6.13.0.sql
@@ -11,3 +11,7 @@ UPDATE pages_fields_translation pft
 INNER JOIN pages p ON (pft.id_pages  = p.id)
 SET content = 'CMS Preferences Update'
 WHERE p.keyword = 'cmsPreferencesUpdate' AND id_languages = 2;
+
+UPDATE `fields`
+SET id_type = get_field_type_id('markdown-inline')
+WHERE `name` = 'label';