+++ server/service/Login.php
@@ -116,8 +116,6 @@ class Login
         // Update the session configuration
         ini_set('session.gc_maxlifetime', SESSION_TIMEOUT);
         ini_set('session.cookie_lifetime', SESSION_TIMEOUT);
-        ini_set('session.gc_probability',1);
-        ini_set('session.gc_divisor',1); 
         session_start();                
         if(!isset($_SESSION['gender'])) $_SESSION['gender'] = MALE_GENDER_ID;
         if(!isset($_SESSION['user_gender'])) $_SESSION['user_gender'] = MALE_GENDER_ID;
@@ -160,7 +158,7 @@ class Login
         }
         // Regenerate session ID to invalidate the old session ID
         session_regenerate_id(true);
-        // session_write_close(); // otherwise it blocks request, check later if session is uesed naywhere else to assgin data.
+        // session_write_close(); // otherwise it blocks request, check later if session is used anywhere else to assign data.
     }
 
     /**