+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### New features
  - rename `Globals` to `Configurations`
  - #478 - add `SESSION_TIMEOUT` in `globals_untracked.php`, the default value is 3600 seconds = 1 hour.
+ - #475 - remove `checkbox` option from `input` style. For the inputs already set to `checkbox` will continue to work.
 
 # v6.12.2
 ### New features